+++ src/api/playlists/handler.js
@@ -12,10 +12,12 @@ class PlaylistsHandler {
     this._validator.validatePlaylistPayload(req.payload);
     const { name } = req.payload;
     const { id: userId } = req.auth.credentials;
+
     const playlist = await this._playlistService.createPlaylist({
       name,
       userId,
     });
+
     return h
       .response({
         status: 'success',
@@ -27,7 +29,9 @@ class PlaylistsHandler {
 
   async getUserPlaylist(req, h) {
     const { id: userId } = req.auth.credentials;
+
     const playlists = await this._playlistService.getPlaylists(userId);
+
     return h.response({
       status: 'success',
       message: 'Playlists retrieved successfully',
@@ -38,10 +42,12 @@ class PlaylistsHandler {
   }
 
   async deleteUserPlaylist(req, h) {
-    const { id: playlistId } = req.params;
+    const { playlistId } = req.params;
     const { id: userId } = req.auth.credentials;
-    await this._playlistService.verifyPlaylistUser(playlistId, userId);
+
+    await this._playlistService.verifyPlaylistOwner(playlistId, userId);
     await this._playlistService.deletePlaylist(playlistId);
+
     return h.response({
       status: 'success',
       message: 'Playlist deleted successfully',
